refactor : 랭킹페이지 코드 압축 / 컴포넌트화  chore : 메인페이지 설문조사 치킨 배너 이미지 주석처리 / 프로필수정 유효성검사 메세지 보더 삭제
