+++ src/pages/Ranking/RankingBox.jsx
@@ -6,21 +6,14 @@ import styles from "./../css/rankingPage.module.css";
 import dropdownBtn from "../../common/svg/dropdown_icon.svg";
 import font from "../../common/css/font.module.css";
 import { useNavigate } from "react-router-dom";
+import RankingCtn from "./RankingCtn";
 
-const RankingBox = ({ btsOn, setBtsOn, ageMode, setAgeMode, type, setType }) => {
+const RankingBox = ({ btsOn, setBtsOn, ageMode, type, setType }) => {
   const dispatch = useDispatch();
   const getMyRanking = useSelector(state => state.ranking.myRanking);
   const getAllRanking = useSelector(state => state.ranking.ranking);
-  const userTimeSet = Math.floor(getMyRanking.savedStudyTime / 1000);
-  const myHour = parseInt(userTimeSet / 3600);
-  const myMinutes = parseInt((userTimeSet % 3600) / 60);
   const navi = useNavigate();
 
-  const agePick = [
-    { ko: "전체 랭킹", en: "all" },
-    { ko: "20대 랭킹", en: "twenty" },
-    { ko: "30대 랭킹", en: "thirty" }
-  ];
   const datePick = [
     { ko: "일간", en: "day" },
     { ko: "주간", en: "week" },
@@ -40,18 +33,9 @@ const RankingBox = ({ btsOn, setBtsOn, ageMode, setAgeMode, type, setType }) => 
     setBtsOn(false);
   };
 
-  console.log(type);
-
   return (
     <>
-      <div
-        style={{
-          display: "flex",
-          flexDirection: "column",
-          justifyContent: "space-between",
-          height: "100%"
-        }}
-      >
+      <div className={styles.container}>
         <div>
           <div className={btsOn ? styles.blurIn : styles.blurOut} onClick={modalOffHandler}>
             <div className={`${styles.rankingType} ${font.subtitle2_600_16}`}>
@@ -70,45 +54,32 @@ const RankingBox = ({ btsOn, setBtsOn, ageMode, setAgeMode, type, setType }) => 
               {datePick.map((day, i) => {
                 return (
                   <div key={i}>
-                    {mode === day.ko ? (
-                      <button
-                        className={`${styles.button} ${font.subtitle4_600_12}`}
-                        style={{
-                          backgroundColor: "var(--neutral-40)",
-                          color: "var(--neutral-100)"
-                        }}
-                        onClick={() => {
-                          setMode(day.ko);
-                          setType({ ...type, period: day.en });
-                        }}
-                      >
-                        {day.ko}
-                      </button>
-                    ) : (
-                      <button
-                        className={`${styles.button} ${font.subtitle4_600_12}`}
-                        style={{
-                          backgroundColor: "var(--neutral-30)",
-                          color: "var(--neutral-70)"
-                        }}
-                        onClick={() => {
-                          setType({ ...type, period: day.en });
-
-                          setMode(day.ko);
-                        }}
-                      >
-                        {day.ko}
-                      </button>
-                    )}
+                    <button
+                      className={`${styles.button} ${font.subtitle4_600_12}`}
+                      style={
+                        mode === day.ko
+                          ? {
+                              backgroundColor: "var(--neutral-40)",
+                              color: "var(--neutral-100)"
+                            }
+                          : {
+                              backgroundColor: "var(--neutral-30)",
+                              color: "var(--neutral-70)"
+                            }
+                      }
+                      onClick={() => {
+                        setMode(day.ko);
+                        setType({ ...type, period: day.en });
+                      }}
+                    >
+                      {day.ko}
+                    </button>
                   </div>
                 );
               })}
             </div>
             <div className={styles.rankingContainer}>
               {getAllRanking.map((rankbox, i) => {
-                const userTimeSet = Math.floor(rankbox.savedStudyTime / 1000);
-                const hour = parseInt(userTimeSet / 3600);
-                const minutes = parseInt((userTimeSet % 3600) / 60);
                 return (
                   <div key={i}>
                     <div
@@ -120,40 +91,7 @@ const RankingBox = ({ btsOn, setBtsOn, ageMode, setAgeMode, type, setType }) => 
                           : `${styles.allStatus} ${font.subtitle4_600_12}`
                       }
                     >
-                      <div className={styles.nameBox}>
-                        <span className={i > 2 ? styles.padding : styles.userRank}>
-                          {i === 0 && "👑"}
-                          {i === 1 && "🥈"}
-                          {i === 2 && "🥉"}
-                          {i + 1}
-                        </span>
-                        <div className={styles.userBox}>
-                          <p className={`${styles.userNickname} ${font.subtitle4_600_12}`}>{rankbox.nickname}</p>
-                          <p
-                            className={
-                              i === 0
-                                ? `${styles.topUserSpec} ${font.caption2_300_10} `
-                                : `${styles.userSpec} ${font.caption2_300_10} `
-                            }
-                          >
-                            {" "}
-                            {rankbox.specialty}
-                          </p>
-                        </div>
-                      </div>
-                      <div className={styles.timeBox}>
-                        <span className={`${styles.userTime} ${font.subtitle4_600_12}`}>
-                          <div className={styles.hour}>{hour < 10 ? "0" + hour : hour}시간</div>
-                          <div className={styles.minutes}> {minutes < 10 ? "0" + minutes : minutes}분</div>
-                        </span>
-                        <div className={styles.dotBox}>
-                          {rankbox.studying ? (
-                            <div className={i === 0 ? styles.rankerDot : styles.greendot}></div>
-                          ) : (
-                            <div className={styles.emptyDot}></div>
-                          )}
-                        </div>
-                      </div>
+                      <RankingCtn i={i} rankbox={rankbox} />
                     </div>
                   </div>
                 );
@@ -163,37 +101,7 @@ const RankingBox = ({ btsOn, setBtsOn, ageMode, setAgeMode, type, setType }) => 
         </div>
         <div style={{ position: "relative" }}>
           <div className={styles.myStatus}>
-            <div className={styles.nameBox}>
-              <span
-                className={
-                  getMyRanking.rank > 2
-                    ? `${styles.padding} ${font.subtitle4_600_12}`
-                    : `${styles.userRank} ${font.subtitle4_600_12}`
-                }
-              >
-                {getMyRanking.rank === 1 && "👑"}
-                {getMyRanking.rank === 2 && "🥈"}
-                {getMyRanking.rank === 3 && "🥉"}
-                {getMyRanking.rank ? getMyRanking.rank : "--"}
-              </span>
-              <div className={styles.userBox}>
-                <p className={`${styles.userNickname} ${font.subtitle4_600_12}`}>{getMyRanking.nickname}</p>
-                <p className={`${styles.userSpec} ${font.caption2_300_10}`}>{getMyRanking.specialty}</p>
-              </div>
-            </div>
-            <div className={styles.timeBox}>
-              <span className={`${styles.userTime} ${font.subtitle4_600_12}`}>
-                <div className={styles.hour}>{myHour < 10 ? "0" + myHour : myHour}시간</div>
-                <div className={styles.minutes}> {myMinutes < 10 ? "0" + myMinutes : myMinutes}분</div>
-              </span>
-              <div className={styles.dotBox}>
-                {getMyRanking.studying ? (
-                  <div className={styles.greendot}></div>
-                ) : (
-                  <div className={styles.emptyDot}></div>
-                )}
-              </div>
-            </div>
+            <RankingCtn rankbox={getMyRanking} i={getMyRanking.rank - 1} myStatus={true} />
           </div>
           <Footer />
         </div>
