+++ src/pages/css/rankingPage.module.css
@@ -9,6 +9,13 @@
   padding: 1rem 1rem;
 }
 
+.container {
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  height: 100%;
+}
+
 .blurIn {
   filter: blur(3px);
   transition: all 0.4s ease-in-out;
@@ -88,6 +95,7 @@
 }
 
 .rankingContainer {
+  border: 1px solid white;
   margin: 0 auto;
   width: 100%;
   padding: 0 16px 16px 16px;
@@ -176,6 +184,7 @@
   flex-direction: row;
   align-items: center;
   padding-left: 1rem;
+  justify-content: space-between;
 }
 
 .topRanker {
@@ -189,6 +198,7 @@
   flex-direction: row;
   align-items: center;
   padding-left: 1rem;
+  justify-content: space-between;
 }
 
 .otherRanker {
@@ -202,6 +212,7 @@
   flex-direction: row;
   align-items: center;
   padding-left: 1rem;
+  justify-content: space-between;
 }
 
 .timeBox {
@@ -209,6 +220,7 @@
   display: flex;
   flex-direction: row;
   align-items: center;
+  margin-right: 1rem;
 }
 
 .dotBox {
@@ -253,6 +265,7 @@
   align-items: center;
   padding-left: 1rem;
   z-index: 10;
+  justify-content: space-between;
 }
 
 .nameBox {
